+++ sdk/apicenter/arm-apicenter/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Release History
     
+## 1.1.0 (2024-03-18)
+    
+**Features**
+
+  - Added operation group DeletedServices
+  - Added Interface DeletedService
+  - Added Interface DeletedServiceListResult
+  - Added Interface DeletedServiceProperties
+  - Added Interface DeletedServicesDeleteOptionalParams
+  - Added Interface DeletedServicesGetHeaders
+  - Added Interface DeletedServicesGetOptionalParams
+  - Added Interface DeletedServicesListNextOptionalParams
+  - Added Interface DeletedServicesListOptionalParams
+  - Added Type Alias DeletedServicesGetResponse
+  - Added Type Alias DeletedServicesListNextResponse
+  - Added Type Alias DeletedServicesListResponse
+    
+    
 ## 1.0.0 (2024-02-21)
 
-The package of @azure/arm-apicenter is using our next generation design principles. To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/js/mgmt/quickstart ).
+The package of @azure/arm-apicenter is using our next generation design principles. To learn more, please refer to our documentation [Quick Start](https://aka.ms/js-track2-quickstart).